KTB bond auction scraper
The auction results for Korean Treasury Bonds (KTB) are not widely accessible on platforms like Bloomberg and the information is very disorganized across the Korea Treasury's English website. 

This scraper uses Selenium and to extract the data for a given range and compiles that KTB auction data data in a CSV which is friendly for analysis for most financial analysts.